+++ NEWS.md
@@ -1,7 +1,12 @@
 # rgl 1.3.25
 
+<<<<<<< HEAD
 * Added the `latex3d()` function to draw LaTeX text using the
 `xdvir` package.
+=======
+* Support for classes defined in the `tripack` package
+has been dropped at the request of CRAN.
+>>>>>>> master
 
 # rgl 1.3.24